id検索(バックエンド)の実装

+++ SpringTransferProject/src/main/java/com/example/transfer/repository/TransferRepository.java
@@ -4,7 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.transfer.entity.TransferRoute;
 
-public interface TransferRepository extends JpaRepository<TransferRoute, Integer> {
-	    // JpaRepositoryを継承することで、基本的なCRUD操作（保存、検索など）が自動的に提供される
+public interface TransferRepository extends JpaRepository<TransferRoute, Long> {
+
+	
+	 // JpaRepositoryを継承することで、基本的なCRUD操作（保存、検索など）が自動的に提供される
 
 }
